pyproject.toml
[project]
name = "leetcode-solver"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "scoopx", email = "hx364@nyu.edu"},
]
dependencies = [
    "requests>=2.32.3",
    "bs4>=0.0.2",
    "google>=3.0.0",
    "google-generativeai>=0.8.3",
    "jinja2>=3.1.5",
    "python-dotenv>=1.0.1",
    "langchain>=0.3.13",
    "langchain-openai>=0.2.14",
    "langchain-anthropic>=0.3.1",
    "langchain-google-genai>=2.0.7",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
